LCD/Keypad code for Wind Tunnel
This code contains the LCD and Keypad code for use with an ATmega328pb microcontroller. This serves as the basis for the Wind Tunnel embedded portion of the project.

Known bugs:
When a number is pressed on the keypad, it may print multiple times.
When "0" is pressed on the keypad it prints several times, stops, and repeats.

To do:
Fix keypad. Add PWM code to send frequency to the motor. Add Enable and CW/CCW modes for motor control. Rewire LCD as necessary to allow appropriate ports to work with PWM.