LCD/Keypad User Interface Code
Known bugs:
When a number is pressed on the keypad, it may print several times to the keypad.
When "0" is pressed on the keypad, it prints several times, stops, then repeats infinitely.

To do:
Fix keypad. Add PWM output to control motor speed. Add output to turn on the Enable on the motor and CW/CCW directions.